refactor(codex): share node catalog record lookup
