+++ extensions/codex/src/session-catalog-terminal.ts
@@ -7,28 +7,20 @@ import type {
   OpenClawPluginApi,
   OpenClawPluginNodeHostCommand,
 } from "openclaw/plugin-sdk/plugin-entry";
-import type { PluginRuntime } from "openclaw/plugin-sdk/plugin-runtime";
 import type { SessionCatalogTerminalPlan } from "openclaw/plugin-sdk/session-catalog";
 import { resolveCodexAppServerLocalHomeDir } from "./app-server/auth-start-options.js";
 import type { resolveCodexSupervisionAppServerRuntimeOptions } from "./app-server/config-runtime.js";
 import type { CodexCatalogHome } from "./session-catalog-homes.js";
+import { lookupNodeCodexCatalogRecord } from "./session-catalog-node-lookup.js";
 import {
   CatalogParamsError,
   CODEX_APP_SERVER_THREADS_CAPABILITY,
   CODEX_APP_SERVER_THREADS_LIST_COMMAND,
   CODEX_LOCAL_SESSION_HOST_ID,
-  CODEX_SESSION_CATALOG_MAX_PAGE_LIMIT,
   isInteractiveThreadSource,
-  MAX_ACTION_CATALOG_PAGES,
-  NODE_INVOKE_TIMEOUT_MS,
-  unwrapNodeInvokePayload,
 } from "./session-catalog-parsing.js";
 import { resolveNodeHostExecutable, runNodePtyCommand } from "./session-catalog-pty.runtime.js";
-import type {
-  CodexSessionCatalogControl,
-  CodexSessionCatalogPage,
-  CodexSessionCatalogSession,
-} from "./session-catalog-types.js";
+import type { CodexSessionCatalogControl } from "./session-catalog-types.js";
 
 export const CODEX_TERMINAL_RESUME_COMMAND = "codex.terminal.resume.v1";
 export const CODEX_TERMINAL_START_COMMAND = "codex.terminal.start.v1";
@@ -189,45 +181,6 @@ export function createCodexTerminalNodeHostCommand(
   };
 }
 
-async function resolveNodeCatalogEligibleThread(params: {
-  agentId: string;
-  runtime: PluginRuntime;
-  nodeId: string;
-  threadId: string;
-  parseCatalogPage: (value: unknown) => CodexSessionCatalogPage;
-}): Promise<CodexSessionCatalogSession> {
-  let cursor: string | undefined;
-  const seenCursors = new Set<string>();
-  for (let pageIndex = 0; pageIndex < MAX_ACTION_CATALOG_PAGES; pageIndex += 1) {
-    const raw = await params.runtime.nodes.invoke({
-      nodeId: params.nodeId,
-      command: CODEX_APP_SERVER_THREADS_LIST_COMMAND,
-      params: {
-        agentId: params.agentId,
-        limit: CODEX_SESSION_CATALOG_MAX_PAGE_LIMIT,
-        ...(cursor ? { cursor } : {}),
-      },
-      timeoutMs: NODE_INVOKE_TIMEOUT_MS,
-      scopes: ["operator.write"],
-    });
-    const page = params.parseCatalogPage(unwrapNodeInvokePayload(raw));
-    const record = page.sessions.find((candidate) => candidate.threadId === params.threadId);
-    if (record) {
-      if (isInteractiveThreadSource(record.source)) {
-        return record;
-      }
-      break;
-    }
-    const nextCursor = page.nextCursor?.trim();
-    if (!nextCursor || seenCursors.has(nextCursor)) {
-      break;
-    }
-    seenCursors.add(nextCursor);
-    cursor = nextCursor;
-  }
-  throw new CatalogParamsError("Codex session is not a non-archived interactive Codex session");
-}
-
 export async function openCodexCatalogTerminal(
   params: {
     agentId: string;
@@ -235,7 +188,6 @@ export async function openCodexCatalogTerminal(
     control: CodexSessionCatalogControl;
     hostId: string;
     threadId: string;
-    parseCatalogPage: (value: unknown) => CodexSessionCatalogPage;
     source?: CodexCatalogHome;
   } & CodexTerminalConfigSources,
 ): Promise<SessionCatalogTerminalPlan> {
@@ -276,13 +228,16 @@ export async function openCodexCatalogTerminal(
   if (!node) {
     throw new CatalogParamsError("paired-node Codex terminal is unavailable");
   }
-  const record = await resolveNodeCatalogEligibleThread({
+  const lookup = await lookupNodeCodexCatalogRecord({
     agentId: params.agentId,
     runtime: params.api.runtime,
     nodeId,
     threadId: params.threadId,
-    parseCatalogPage: params.parseCatalogPage,
   });
+  if (lookup.kind !== "found" || !isInteractiveThreadSource(lookup.record.source)) {
+    throw new CatalogParamsError("Codex session is not a non-archived interactive Codex session");
+  }
+  const record = lookup.record;
   return {
     kind: "node",
     nodeId,
